rename package
remove leftovers
