+++ pkg/ac/ac_suite_test.go
@@ -4,7 +4,7 @@ SPDX-FileCopyrightText: 2022 SAP SE or an SAP affiliate company and Gardener con
 SPDX-License-Identifier: Apache-2.0
 */
 
-package acc_test
+package ac_test
 
 import (
 	"testing"
