+++ pkg/ac/access_restriction.go
@@ -3,7 +3,7 @@ SPDX-FileCopyrightText: 2022 SAP SE or an SAP affiliate company and Gardener con
 
 SPDX-License-Identifier: Apache-2.0
 */
-package acc
+package ac
 
 import (
 	"context"
