+++ pkg/config/config.go
@@ -18,7 +18,7 @@ import (
 	"k8s.io/client-go/tools/clientcmd"
 	clientcmdapi "k8s.io/client-go/tools/clientcmd/api"
 
-	"github.com/gardener/gardenctl-v2/pkg/acc"
+	"github.com/gardener/gardenctl-v2/pkg/ac"
 )
 
 // Config holds the gardenctl configuration
@@ -46,7 +46,7 @@ type Garden struct {
 	// +optional
 	Patterns []string `yaml:"patterns,omitempty" json:"patterns,omitempty"`
 	// AccessRestrictions is a list of access restriction definitions
-	AccessRestrictions []acc.AccessRestriction `yaml:"accessRestrictions,omitempty" json:"accessRestrictions,omitempty"`
+	AccessRestrictions []ac.AccessRestriction `yaml:"accessRestrictions,omitempty" json:"accessRestrictions,omitempty"`
 }
 
 // LoadFromFile parses a gardenctl config file and returns a Config struct
